# src/main/resources/Documentation/Swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AR Backend API",
    "description": "OpenAPI documentation for the Multi-Tenant AR Backend",
    "contact": {
      "name": "Lucas",
      "url": "https://your-website.com",
      "email": "jclucasasp@gmail.com"
    },
    "license": {
      "name": "Proprietary",
      "url": "https://your-website.com/license"
    },
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "SessionAuth": []
    },
    {
      "X-API-KEY": []
    }
  ],
...
  "paths": {
...
    "/api/v1/discussions/questions": {
      "post": {
        "tags": [
          "Course Discussions"
        ],
        "summary": "Ask a question",
        "description": "Allows a student to post a new question to a course section.",
        "operationId": "askQuestion",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CourseQuestionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CourseQuestionResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/discussions/questions/{questionId}/replies": {
      "post": {
        "tags": [
          "Course Discussions"
        ],
        "summary": "Post a reply",
        "description": "Allows students or staff to reply to a specific question.",
        "operationId": "postReply",
        "parameters": [
          {
            "name": "questionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReplyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ReplyResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/discussions/course/{courseId}": {
      "get": {
        "tags": [
          "Course Discussions"
        ],
        "summary": "Get all questions for a course",
        "description": "Returns a list of questions filtered by course and organisation.",
        "operationId": "getQuestions",
        "parameters": [
          {
            "name": "courseId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CourseQuestionResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CourseQuestionRequest": {
        "type": "object",
        "properties": {
          "courseId": {
            "type": "integer",
            "format": "int64"
          },
          "sectionId": {
            "type": "integer",
            "format": "int64"
          },
          "title": {
            "type": "string"
          },
          "body": {
            "type": "string"
          }
        },
        "required": [
          "body",
          "title"
        ]
      },
      "CourseQuestionResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "title": {
            "type": "string"
          },
          "body": {
            "type": "string"
          },
          "studentName": {
            "type": "string"
          },
          "isResolved": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "replies": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReplyResponse"
            }
          }
        }
      },
      "ReplyResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "body": {
            "type": "string"
          },
          "authorName": {
            "type": "string"
          },
          "isStaff": {
            "type": "boolean"
          },
          "isAccepted": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ReplyRequest": {
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "SessionAuth": {
        "type": "apiKey",
        "description": "Authentication via JSESSIONID cookie after login",
        "name": "JSESSIONID",
        "in": "cookie"
      },
      "X-API-KEY": {
        "type": "apiKey",
        "description": "API auth learningObjectives for students",
        "name": "X-API-KEY",
        "in": "header"
      }
    }
  }
}